更新日誌及多項代碼改進
- 在 `CHANGELOG.md` 中新增 2025 年 1 月 8 日的更新記錄，提到新增 CXXStack 的栈清空函數。
- 在 `unzip.cpp` 中添加 `__fallthrough;` 以標示 switch-case 的自動跳轉。
- 修改 `zip.cpp` 中 `FormatZipMessageZ` 函數的參數類型以支持 Unicode。
- 更新 `PULL_REQUEST_TEMPLATE.zh-CN.md` 中的命名規則，強調代碼架構和內存管理。
- 在 `cxxstack.h` 中將 `useLock` 變數改名為 `useLock_`，並新增 `clear` 函數。
- 調整 `cxxcrypto.cpp` 中加密和解密函數的參數類型，使用 `static_cast` 進行明確類型轉換。
- 重構 `test.cpp` 中的測試類，新增多個針對 `CXXString` 的測試方法。
- 在 `ConsoleTest.vcxproj` 中添加 `unzip.cpp` 和 `zip.cpp`，並將子系統改為 Console。
- 移除 `main.cpp` 中對 `CXXZip::compress` 和 `CXXZip::uncompress` 的調用以簡化邏輯。
- 更新 `module_cxxcore.h` 的標頭信息，記錄 CXXCrypto 和 CXXStack 的變更。

+++ CXXCore-master/src/cxxcrypto.cpp
@@ -468,7 +468,7 @@ bool CXXCrypto::decryptCBC(const CXXString& input)
 		if (!d_ptr->outputBuffer_.empty())
 		{
 			size_t paddingSize = d_ptr->outputBuffer_.back();
-			if (paddingSize <= d_ptr->outputBuffer_.size())
+			if (paddingSize <= static_cast<size_t>(d_ptr->outputBuffer_.size()))
 			{
 				d_ptr->outputBuffer_.resize(d_ptr->outputBuffer_.size() - paddingSize);
 			}
@@ -927,7 +927,7 @@ bool CXXCrypto::encryptDataCBC(const BYTE* data, __int64 dataSize, CXXByteArray*
 
 		// Get required size for encrypted data
 		DWORD requiredSize = 0;
-		if (!CXXCryptoPrivate::encrypt(hKey, paddedData.data(), paddedData.size(),
+		if (!CXXCryptoPrivate::encrypt(hKey, paddedData.data(), static_cast<ULONG>(paddedData.size()),
 			nullptr, 0, requiredSize, 0))
 		{
 			break;
@@ -940,11 +940,11 @@ bool CXXCrypto::encryptDataCBC(const BYTE* data, __int64 dataSize, CXXByteArray*
 		BCRYPT_AUTHENTICATED_CIPHER_MODE_INFO authInfo;
 		BCRYPT_INIT_AUTH_MODE_INFO(authInfo);
 		authInfo.pbNonce = const_cast<BYTE*>(d_ptr->getIV().data());
-		authInfo.cbNonce = d_ptr->getIV().size();
+		authInfo.cbNonce = static_cast<ULONG>(d_ptr->getIV().size());
 
 		DWORD encryptedSize = 0;
-		NTSTATUS status = BCryptEncrypt(hKey, paddedData.data(), paddedData.size(),
-			&authInfo, nullptr, 0, output->data(), output->size(),
+		NTSTATUS status = BCryptEncrypt(hKey, paddedData.data(), static_cast<ULONG>(paddedData.size()),
+			&authInfo, nullptr, 0, output->data(), static_cast<ULONG>(output->size()),
 			&encryptedSize, 0);
 
 		if (!BCRYPT_SUCCESS(status))
@@ -1008,11 +1008,11 @@ bool CXXCrypto::decryptDataCBC(const BYTE* data, __int64 dataSize, CXXByteArray*
 		BCRYPT_AUTHENTICATED_CIPHER_MODE_INFO authInfo;
 		BCRYPT_INIT_AUTH_MODE_INFO(authInfo);
 		authInfo.pbNonce = const_cast<BYTE*>(d_ptr->getIV().data());
-		authInfo.cbNonce = d_ptr->getIV().size();
+		authInfo.cbNonce = static_cast<ULONG>(d_ptr->getIV().size());
 
 		DWORD decryptedSize = 0;
-		NTSTATUS status = BCryptDecrypt(hKey, const_cast<BYTE*>(data), dataSize,
-			&authInfo, nullptr, 0, output->data(), output->size(),
+		NTSTATUS status = BCryptDecrypt(hKey, const_cast<BYTE*>(data), static_cast<ULONG>(dataSize),
+			&authInfo, nullptr, 0, output->data(), static_cast<ULONG>(output->size()),
 			&decryptedSize, 0);
 
 		if (!BCRYPT_SUCCESS(status))
